Create readfileparsevalue
This program prompts the user for the name of file to read, then the lines that does not match with "X-DSPAM-Confidence:" are excluded. Also, the number of lines that match "X-DSPAM-Confidence:" are counted. These lines contain a float number in the end. Then we convert the string from these lines to 'floats', removing the beginning "X-DSPAM-Confidence:". Finally, we make the sum of the numbers and divide by the number of lines to get the average.